chore: 모든 엔티티 필드 설계 완료

+++ Back/src/main/java/com/kkosunnae/deryeogage/domain/simulation/SimulationEntity.java
@@ -58,10 +58,10 @@ public class SimulationEntity {
     private Float train4;
 
     @Column
-    private Integer health;
+    private Byte health;
 
     @Column(name = "quiz_num")
-    private Integer quizNum;
+    private Byte quizNum;
 
     @Column(length = 20)
     private String title;
